evals/requirements.txt
accelerate
transformers
wget
evaluate>=0.3.0
huggingface_hub>=0.11.1
datasets>=2.14.6
termcolor
appdirs
tempdir
multipledispatch
numpy
tqdm
fire
rich
tree_sitter>=0.22.0
tree-sitter-python
psutil
antlr4-python3-runtime==4.11.0
ai2-tango
torchmetrics==0.11.1
more_itertools
spacy>=3.0.0
bettermap
bert_score            
sacrebleu>=1.5.0
scikit-learn>=0.24.1
iso639>=0.1.4
rouge-score>=0.0.4
mypy_extensions
pytest
jinja2
pyyaml>=5
sentencepiece
mosestokenizer>=1.0.0
colorama
gitpython
pytablewriter
aenum==3.1.15
pycountry
nltk==3.8.1
gradio-client
langdetect
immutabledict
jsonlines
litellm
alpaca_eval
boto3
smart-open
pygsheets
wandb
